moved classes to model.py
also addressing jonathan's concerns, removing card counts and making a function to compute them in the classes
